test: migrate TransactionMachine tests to AVA
