+++ jest2ava.sh
@@ -21,7 +21,9 @@ sed -r -i 's#expect\((.+?)\).not.toInclude\((.+?)\)#t.false\(\1.includes\(\2\)\)
 
 sed -r -i 's#assertZWaveError\(#assertZWaveErrorAva\(t, #' "$1"
 sed -r -i 's#assertCC\(#assertCCAva\(t, #' "$1"
+
 sed -r -i 's#jest\.fn\(\)#sinon.spy\(\)#' "$1"
+sed -r -i 's#jest\.Mock#sinon.SinonStub#' "$1"
 sed -r -i 's#expect\((.+?)\).toBeCalled\(\)#sinon.assert.called\(\1\)#' "$1"
 sed -r -i 's#expect\((.+?)\).not.toBeCalled\(\)#sinon.assert.notCalled\(\1\)#' "$1"
 sed -r -i 's#expect\((.+?)\).toBeCalledTimes\((.+?)\)#t.is\(\1.callCount, \2\)#' "$1"
@@ -33,3 +35,4 @@ sed -r -i 's#expect\((.+?)\).toBeInstanceOf\((.+?)\)#t.true\(\1 instanceof \2\)#
 sed -r -i 's#expect\((.+?)\).toMatchObject\(#t.like\(\1, #' "$1"
 
 sed -r -i 's#expect\((.+?)\).toResolve\(\)#t.notThrowsAsync\(\(\) => \1\)#' "$1"
+
